Moved BACK_URL to .env file
