+++ app/src/constant/index.ts
@@ -1,1 +1,2 @@
 export const TIPS_ADDRESS = import.meta.env.VITE_APP_TIPS_ADDRESS
+export const BACKEND_URL = import.meta.env.VITE_APP_BACKEND_URL
