--- app/src/utils/util.ts
@@ -1,1 +1,0 @@
-export const BACKEND_URL = 'http://localhost:8082'
